requirements.txt
python-telegram-bot >=20.0
python-dotenv >=0.16.0
babel >=2.7.0
apscheduler >=3.6.3
requests

-r lib/api/requirements.txt